+++ src/System.IO.FileSystem/tests/System.IO.FileSystem.Tests.csproj
@@ -16,7 +16,6 @@
   <ItemGroup>
     <Compile Include="FileInfo\IsReadOnly.cs" />
     <Compile Include="FileInfo\Replace.cs" />
-    <Compile Include="FileStream\Constructor.cs" />
     <Compile Include="FileStream\Handle.cs" />
     <Compile Include="Directory\GetLogicalDrives.cs" />
     <Compile Include="FileStream\EndRead.cs" />
